Add log ID to TransparencyLogEntry response
Fixes #285

To minimize the transition for clients between Rekor v1 and v2, we will
include the log ID in the response. This log ID is used by clients to
look up the correct instance in a trust root, though some clients ignore
it and build a pool of verification keys instead, and compare a
truncated log ID to the checkpoint signature key prefix.

I've chosen to make the log ID equal to the non-truncated checkpoint key
ID. For clients that treat the log ID as an opaque string when looking
up the instance in the trusted root, no changes should be needed.
Similarly, if a client uses the log ID to compare against the checkpoint
key ID, like what sigstore-python does, no changes are needed.

In the TransparencyLogInstance in the trusted root, both the log ID and
checkpoint key ID will be set to the same value. Longer-term, we'll
deprecate log ID. Even longer-term, ideally clients would enumerate all
keys and compute the checkpoint key IDs themselves, but for now, we'll
keep populating the IDs in the trusted root for convenience.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>