Return 405 Method Not Allowed for read APIs
When we were deploying the service to staging, we noticed that the read
APIs were returning 200 with an empty body for some reason. The root
cause is due to how we handle returning an Unimplemented gRPC status
code.

In order to return the non-retryable 405 Method Not Allowed when the
server goes into a read-only mode, we added special handling when the
gRPC status code is set to Unimplemented. Otherwise, it returns a 501
Not Implemented, which is both retryable and not an accurate status for
when the server is read-only.

When the status code is Unimplemented, we set the response code and body
in the httpResponseModifier function based on internally-set gRPC
headers. This worked correctly for the read-only response. However, for
the 3 unimplemented read APIs, while we were setting a gRPC response
code of Unimplemented, we weren't setting the internal headers. This
meant that we skipped over where the HTTP response code/body is set, and
so we returned a default 200 with an empty body for the read APIs.

This PR adds the internal gRPC headers and a test that we are returning
405 as expected.

I've confirmed that 405 is an appropriate response code. Besides being
non-retryable, the RFC notes that servers should always implement GET
methods, so a 405 is most appropriate -
https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/501

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>